+++ trailmet/utils/functions.py
@@ -1,0 +1,14 @@
+import random
+import os
+import numpy as np
+import torch
+
+def seed_everything(seed):
+    "sets the random seed to ensure reproducibility"
+    random.seed(seed)
+    os.environ['PYTHONHASHSEED'] = str(seed)
+    np.random.seed(seed)
+    torch.manual_seed(seed)
+    torch.cuda.manual_seed(seed)
+    torch.cuda.manual_seed_all(seed)
+    torch.backends.cudnn.deterministic = True
